pyproject.toml
[tool.poetry]
name = "arielBot"
version = "0.3.8"
description = ""
authors = ["Polyisoprene <lzxder@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.3.1"
click = "^8.1.7"
nonebot-adapter-onebot = "^2.4.3"
fastapi = "^0.111.0"
aiosqlite = "^0.21.0"
pycryptodome = "^3.21.0"
qrcode = "^8.0"
pypng = "^0.20220715.0"
dynrender-skia = "^0.2.9"
nonebot-plugin-apscheduler = "^0.5.0"

[tool.poetry.group.dev.dependencies]

[tool.poetry.scripts]
ariel = "arielbot.__init__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

